get_games/requirements.txt
numpy
requests
python-dotenv
tqdm
pandas